+++ index.js
@@ -98,7 +98,7 @@ var command = cli.create('jora', '[query]')
     .option('-q, --query <query>', 'Jora query')
     .option('-i, --input <filename>', 'Input file')
     .option('-o, --output <filename>', 'Output file (outputs to stdout if not set)')
-    .option('-p, --pretty [indent]', 'Pretty print with optionally specified indentation(4 spaces by default)', value =>
+    .option('-p, --pretty [indent]', 'Pretty print with optionally specified indentation (4 spaces by default)', value =>
         value === undefined ? 4 : Number(value) || false
     , false)
     .action(function(args) {