+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.1 (July 4, 2019)
 
-- Used dist version of `jora` to reduce startup time (up to 10x)
+- Used dist version of `jora` to reduce startup time (up to 10x times)
 - Removed wrongly added dependency
 - Fixed command name in help info
 - Fixed returning `undefined` when no query, now command returns input itself as when query is empty string