+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-## next
+## 1.1.0 (August 22, 2019)
 
-- Added colored output when appropriate. Use `--no-color` option to suppress color output (#2)
-- Fixed output `undefined` value
+- Added output highlighting when appropriate. Use `--no-color` option to suppress color output (#2)
+- Fixed output of `undefined` value
 
 ## 1.0.1 (July 4, 2019)